+++ deepparse/dataset_container/dataset_container.py
@@ -78,7 +78,6 @@ class DatasetContainer(Dataset, ABC):
             raise DataError("Some addresses data points are empty.")
 
         if validate_if_any_whitespace_only(string_elements=data_to_validate):
-            self.data_tags_not_the_same_len_diff()
             raise DataError("Some addresses only include whitespace thus cannot be parsed.")
 
     def _data_is_a_list(self):
@@ -96,6 +95,11 @@ class DatasetContainer(Dataset, ABC):
             raise DataError("Some tags data points are empty.")
 
         if not self._data_tags_is_same_len_then_address():
+            print(
+                f"Some addresses (whitespace-split) and the tags associated with them are not the same len. Here is"
+                f"the report of those cases where len differ to help you out:\n"
+                f"{self._data_tags_not_the_same_len_diff()}"
+            )
             raise DataError("Some addresses (whitespace-split) and the tags associated with them are not the same len.")
 
     def _data_is_list_of_tuple(self) -> bool:
@@ -119,12 +123,16 @@ class DatasetContainer(Dataset, ABC):
     def is_a_train_container(self) -> bool:
         return self.is_training_container
 
-    def data_tags_not_the_same_len_diff(self) -> None:
+    def _data_tags_not_the_same_len_diff(self) -> str:
         diff_indexes = [index for index, data in enumerate(self.data) if len(data[0].split(" ")) != len(data[1])]
+        report = ""
         for diff_index in diff_indexes:
-            print(f"The data point: {self.data[diff_index]}"
-                  f"\n\tLen of the address: {len(self.data[diff_index][0].split(' '))}"
-                  f"\n\tLen of the tags: {len(self.data[diff_index][1])}")
+            report += (
+                f"The data point (with index: {diff_index}: {self.data[diff_index]}"
+                f"\n\tLen of the address: {len(self.data[diff_index][0].split(' '))}"
+                f"\n\tLen of the tags: {len(self.data[diff_index][1])}\n"
+            )
+        return report
 
 
 class PickleDatasetContainer(DatasetContainer):
@@ -220,13 +228,13 @@ class CSVDatasetContainer(DatasetContainer):
     """
 
     def __init__(
-            self,
-            data_path: str,
-            column_names: Union[List, str],
-            is_training_container: bool = True,
-            separator: str = "\t",
-            tag_seperator_reformat_fn: Union[None, Callable] = None,
-            csv_reader_kwargs: Union[None, Dict] = None,
+        self,
+        data_path: str,
+        column_names: Union[List, str],
+        is_training_container: bool = True,
+        separator: str = "\t",
+        tag_seperator_reformat_fn: Union[None, Callable] = None,
+        csv_reader_kwargs: Union[None, Dict] = None,
     ) -> None:
         super().__init__(is_training_container=is_training_container)
         if is_training_container: