+++ tests/dataset_container/test_tools.py
@@ -1,10 +1,10 @@
 from unittest import TestCase
 
-from deepparse.dataset_container import former_python_list, comma_separated_list_reformat
+from deepparse.dataset_container import former_python_list
 
 
 class DatasetContainerToolsTests(TestCase):
-    def test_given_a_str_list_when_parse_return_properly_parse_list(self):
+    def test_given_a_str_list_split_comma_and_whitespace_when_parse_return_properly_parse_list(self):
         a_list = [0, 1, 2, 3]
         str_list = str(a_list)
         expected_parsing = [str(el) for el in a_list]
@@ -12,9 +12,10 @@ class DatasetContainerToolsTests(TestCase):
         actual_parsing = former_python_list(str_list)
         self.assertEqual(expected_parsing, actual_parsing)
 
-    def test_given_a_comma_separated_str_when_parse_return_properly_parse_list(self):
-        a_comma_separated_list = "1, 2, 3, 4"
-        expected_parsing = ["1", "2", "3", "4"]
+    def test_given_a_str_list_split_comma_when_parse_return_properly_parse_list(self):
+        a_list = [0, 1, 2, 3]
+        str_list = "[0,1,2,3]"
+        expected_parsing = [str(el) for el in a_list]
 
-        actual_parsing = comma_separated_list_reformat(a_comma_separated_list)
+        actual_parsing = former_python_list(str_list)
         self.assertEqual(expected_parsing, actual_parsing)